Creates new data function for combining notes and posts
